pyproject.toml
[project]
name = "jobbot"
version = "0.1.0"
description = "Personal automated job-search pipeline: scrape, score, tailor CV/CL, auto-apply, email digest."
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "selectolax>=0.3.21",
  "feedparser>=6.0",
  "playwright>=1.46",
  "anthropic>=0.40",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "tenacity>=8.5",
  "python-dotenv>=1.0",
  "structlog>=24.1",
  "rich>=13.7",
  "weasyprint>=62.3",         # md→pdf for upload only
  "markdown-it-py>=3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "respx>=0.21",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.scripts]
jobbot = "jobbot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jobbot"]

[tool.ruff]
line-length = 100
target-version = "py311"
